update personal info test with user profile
