+++ js/apps/account-ui/test/personal-info/user-profile.json
@@ -1,0 +1,233 @@
+{
+  "attributes": [
+    {
+      "name": "username",
+      "displayName": "${username}",
+      "permissions": {
+        "view": [
+          "admin",
+          "user"
+        ],
+        "edit": [
+          "admin",
+          "user"
+        ]
+      },
+      "validations": {
+        "length": {
+          "min": 3,
+          "max": 255
+        },
+        "username-prohibited-characters": {},
+        "up-username-not-idn-homograph": {}
+      }
+    },
+    {
+      "name": "email",
+      "displayName": "${email}",
+      "required": {
+        "roles": [
+          "user"
+        ]
+      },
+      "permissions": {
+        "view": [
+          "admin",
+          "user"
+        ],
+        "edit": [
+          "admin",
+          "user"
+        ]
+      },
+      "validations": {
+        "email": {},
+        "length": {
+          "max": 255
+        }
+      }
+    },
+    {
+      "name": "firstName",
+      "displayName": "${firstName}",
+      "required": {
+        "roles": [
+          "user"
+        ]
+      },
+      "permissions": {
+        "view": [
+          "admin",
+          "user"
+        ],
+        "edit": [
+          "admin",
+          "user"
+        ]
+      },
+      "validations": {
+        "length": {
+          "max": 255
+        },
+        "person-name-prohibited-characters": {}
+      }
+    },
+    {
+      "name": "lastName",
+      "displayName": "${lastName}",
+      "required": {
+        "roles": [
+          "user"
+        ]
+      },
+      "permissions": {
+        "view": [
+          "admin",
+          "user"
+        ],
+        "edit": [
+          "admin",
+          "user"
+        ]
+      },
+      "validations": {
+        "length": {
+          "max": 255
+        },
+        "person-name-prohibited-characters": {}
+      }
+    },
+    {
+      "name": "select",
+      "displayName": "Select",
+      "selector": {
+        "scopes": [
+          "roles",
+          "offline_access",
+          "role_list",
+          "acr",
+          "address",
+          "microprofile-jwt",
+          "web-origins",
+          "profile",
+          "phone",
+          "email"
+        ]
+      },
+      "permissions": {
+        "edit": [
+          "user"
+        ],
+        "view": [
+          "user"
+        ]
+      },
+      "annotations": {
+        "inputType": "select",
+        "inputHelperTextBefore": "This is helping a lot"
+      },
+      "validations": {
+        "options": {
+          "options": [
+            "one",
+            "two",
+            "three"
+          ]
+        }
+      }
+    },
+    {
+      "name": "hobbies",
+      "displayName": "",
+      "selector": {
+        "scopes": [
+          "roles",
+          "offline_access",
+          "role_list",
+          "acr",
+          "address",
+          "microprofile-jwt",
+          "web-origins",
+          "profile",
+          "phone",
+          "email"
+        ]
+      },
+      "permissions": {
+        "edit": [
+          "user",
+          "admin"
+        ],
+        "view": [
+          "user",
+          "admin"
+        ]
+      },
+      "annotations": {
+        "inputType": "textarea"
+      },
+      "validations": {},
+      "group": "group"
+    },
+    {
+      "name": "email2",
+      "displayName": "Alternative email",
+      "required": {
+        "roles": [
+          "user"
+        ],
+        "scopes": [
+          "roles",
+          "offline_access",
+          "role_list",
+          "acr",
+          "address",
+          "microprofile-jwt",
+          "web-origins",
+          "profile",
+          "phone",
+          "email"
+        ]
+      },
+      "selector": {
+        "scopes": [
+          "roles",
+          "offline_access",
+          "role_list",
+          "acr",
+          "address",
+          "microprofile-jwt",
+          "web-origins",
+          "profile",
+          "phone",
+          "email"
+        ]
+      },
+      "permissions": {
+        "edit": [
+          "admin",
+          "user"
+        ],
+        "view": [
+          "admin",
+          "user"
+        ]
+      },
+      "annotations": {
+        "inputType": "html5-email"
+      },
+      "validations": {
+        "email": {}
+      },
+      "group": "group"
+    }
+  ],
+  "groups": [
+    {
+      "annotations": {},
+      "displayDescription": "We want to be able to tell what kind of person you are please fill this out",
+      "displayHeader": "The Group",
+      "name": "group"
+    }
+  ]
+}
